feat: draft yaml loader function and test harmonizing shared models
moving in the direction of yaml configuration of the entire process, this commit adds a function to
load dataset-specific and shared model information from yaml configuration files. the function is
not yet called in main execution. the basic method functionality is tested but more tests should be
added for resolving various kinds of corner cases.