feat: apply per-variable bounds; refactor data structure that holds per-variable content
Changed around variable.summary data structure to have one more layer of lists, so that variable
content is stored in a list called variables.  Created a new yaml section called globals for content
that's not specific to one variable; updated the data structure to mimic this and have a list called
globals at the same level as the variables list.  Updated all requisite functions and tests.