feat: add experimental support for self-reported ancestry harmonization
each dataset comes with a self-reported ancestry variable, which is super important but currently
unusable due to a complete lack of harmonization. this patch adds some functions that purport to do
this, by attempting partial matching against the stock list of ancestries being used for the v2
questionnaire (it's actually from wikipedia apparently). the functions are not currently deployed in
the primary code, and test support is incomplete, though it does seem to be working more or less as
initially intended. first, string cleaned self-reports are scanned against all valid ancestry tags,
and identical matches are automatically accepted. then, remaining self-reports without perfect
matches are scanned for string similarity against all ancestry tags, and a simple criterion for
discerning between similar matches is applied. this will inevitably need tweaking at the least; we
should prefer no-calls to false calls in the end.