feat: add more tracking to ancestry matching; improve and expand tests
attempted matching of self-reported ancestry to known labels now emits verbose tracking information,
and that information is injected into the report in place of the standard factor non-compliant level
reporting. tests are adjusted to reflect this expanded output format. also added a test for the
top-level caller, which is basically a slightly modified version of the test for the by-vector
harmonize function. still needs to be called from the main code, and the report itself needs
probable correction and inspection.