fix: harmonize consent IDs with input string processing
a minimal version of the standard internal string processing is now applied to consent IDs, to try
to guarantee they are harmonized when the comparison actually occurs, and so the user can specify
raw IDs in all instances. not all processing steps are applied (e.g. excel error code flagging,
Unicode processing); this may need to be revisited. this patch also updates the readme to reflect
the new permitted yaml tags under globals.