fix: pass on unrecognized type entries and crash if the final result is inconsistent with input
this is brought about by some magic invisible characters inside the type cells causing issues. at
least in the instance that this was detected, passing over that instance (with a warning, for
safety) was correct. the sanity check at the end should be conclusive, and should help in situations
when the output from SurveyCTO gains unexpected extra metadata columns