fix: override default intermediates and knit location for concurrent tasks
multiple concurrent report generation tasks were causing race condition issues due to messy cleanup
of tmp files by rmd. this is now overridden to point to tempdir(), whatever that may be set to for a
given task. in some cases, this may also not be desirable; it may be necessary to expose this to
userspace in a future patch.